chore(server/db): 新增bigint类型并更新缓存表version字段类型
1.  在PrimitiveType类型中新增bigint类型支持
2.  将cache_info表的version字段类型从integer改为bigint

+++ server/src/db/dsm/schemas/complete-schema.ts
@@ -593,7 +593,7 @@ export const COMPLETE_SCHEMA: DatabaseSchema = {
         { name: 'id', type: 'id', primaryKey: true, autoIncrement: true },
         { name: 'user_id', type: 'integer', nullable: false },
         { name: 'cache_type', type: 'string', length: 50, nullable: false },
-        { name: 'version', type: 'integer', nullable: false },
+        { name: 'version', type: 'bigint', nullable: false },
         { name: 'cache_data', type: 'text', nullable: false },
         { name: 'expires_at', type: 'datetime', nullable: true },
         { name: 'created_at', type: 'datetime', defaultValue: 'CURRENT_TIMESTAMP' },
